requirements.txt
# Core dependencies
hydra-core==1.3.2
omegaconf==2.3.0
fire==0.5.0

# Elasticsearch
elasticsearch==8.11.1
elasticsearch-dsl==8.11.0

# Data processing
datasets==2.14.6
pyarrow==14.0.1

# NLP preprocessing
nltk==3.8.1

# Utilities
tqdm==4.66.1
python-dotenv==1.0.0

# Performance monitoring
psutil==5.9.6
memory-profiler==0.61.0

# Plotting
matplotlib==3.8.2
seaborn==0.13.0

# Data handling
pandas==2.1.3
numpy==1.24.4

# Testing (optional)
pytest==7.4.3
pytest-cov==4.1.0

# Logging
colorlog==6.8.0

python-dotenv==1.0.0